% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_skratt.R
\name{random_skratt}
\alias{random_skratt}
\title{Generate "skratt" in a random language}
\usage{
random_skratt()
}
\value{
A character vector of a randomly generated version of "skratt".
}
\description{
Generates the Swedish word "skratt" (laugh) in
a random language.
}
\examples{

random_skratt()

}
