% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_skratt.R
\name{specific_skratt}
\alias{specific_skratt}
\title{Generate "skratt" in a specific language}
\usage{
random_skratt()
}
\arguments{
\item{language}{Any of the two letter codes listed in skrattdat$ccode.}
}
\value{
A character vector of a specific version of "skratt".
}
\description{
Generates the Swedish word "skratt" (laugh) in
the specified language.
}
\examples{

specific_skratt()

}
